+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.1.0
 
-- Initial release.
+- Initial release including the `Pubspec` class and a top-level `parsePubspec`
+  function.
+